Flyttade AppState till GUI, la på ObservableObject,  fileName är [ObservableProperty] och gjorde AppState som en [ObservableProperty] i MainViewModel.
Nu uppdateras FileName i MainPage när man laddar en ny fil.
