+++ RealEstate/ViewModels/PaymentViewModel.cs
@@ -10,7 +10,7 @@ using RealEstate.Core.Services;
 using RealEstate.Windows;
 using System.Collections.ObjectModel;
 using System.Windows;
-
+using RealEstate.Models;
 namespace RealEstate.ViewModels;
 
 public partial class PaymentViewModel : ObservableObject, INavigationAware
